Rename to be consistent with other repositories
